+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 
+## [1.2.16] - Unreleased
+- Allow scraping examples for docs.rs.
+- Expose WidgetExt::raw_user_data() and set_raw_user_data().
+
 ## [1.2.15] - 2021-11-11
 - Fix android build (break introduced with version 1.2.11). Thanks @Ar37-rs. 
 - Don't assert that window is shown with set_opacity.